Rename: 폴더명 변경 src/api -> src/service, 폴더명 변경으로 인한 import 수정
