+++ src/types/next-auth.d.ts
@@ -1,0 +1,17 @@
+import { ArtistData } from "@/components/ArtistWrap";
+import NextAuth from "next-auth";
+
+declare module "next-auth" {
+  /**
+   * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
+   */
+  interface Session {
+    user: {
+      /** The user's postal address. */
+      email: string;
+      name: string;
+      profile: string;
+      address: string;
+    };
+  }
+}
